Separate RequiresDefaultInjection from DefaultInjectionCondition
