+++ src/test/kotlin/io/github/freya022/botcommands/test/services/NamedService.kt
@@ -2,7 +2,7 @@ package io.github.freya022.botcommands.test.services
 
 import io.github.freya022.botcommands.api.core.service.annotations.BService
 import io.github.freya022.botcommands.api.core.service.annotations.InterfacedService
-import io.github.freya022.botcommands.internal.core.service.RequiresDefaultInjection
+import io.github.freya022.botcommands.internal.core.service.annotations.RequiresDefaultInjection
 
 @InterfacedService(acceptMultiple = true)
 interface INamedService
