+++ src/test/kotlin/io/github/freya022/botcommands/test/services/NonUniqueProviderTest.kt
@@ -3,7 +3,7 @@ package io.github.freya022.botcommands.test.services
 import io.github.freya022.botcommands.api.core.service.annotations.BConfiguration
 import io.github.freya022.botcommands.api.core.service.annotations.BService
 import io.github.freya022.botcommands.api.core.service.annotations.Lazy
-import io.github.freya022.botcommands.internal.core.service.RequiresDefaultInjection
+import io.github.freya022.botcommands.internal.core.service.annotations.RequiresDefaultInjection
 
 // Even if this is lazy, this should throw as Service1 has multiple definitions and no name matches
 @Lazy
